requirements.txt
Django==2.2.22
django-tinymce==2.6.0
mysqlclient==2.0.3
cryptography==2.6.1
python-decouple==3.2